requirements.txt
./packages/eccn
./packages/scikit-multilearn
numpy==2.0.1
pandas==2.2.2
matplotlib==3.9.1.post1
loky==3.4.1
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
matlabengine==24.2.1
pre-commit
